Added: Room Traverse, see full update
Please change parameters of ARoom directly from the actor in the scene now

ToDo:
Please, use references, auto&, auto* to simplify the reading
Please, use Unreal Docs to find the path of the library you plan to include, those are consistent between systems.
When creating new AActor, untick if it is not something movable or controllable.
It is better to avoid AActor at all and use a custom class for calculating all of these things and THEN spawn rooms in the positions
Also, we need to restretch tiles, the new mesh is coming in the next update.